addin the custom use functianlity and reducing the useefect complex logic in to another file
