+++ src/Movie.jsx
@@ -1,5 +1,6 @@
 import { useEffect, useRef, useState } from "react";
 import Star from "./Star";
+import { useMovies } from "./useMovie";
 
 const tempMovieData = [
   {
@@ -50,16 +51,13 @@ const tempWatchedData = [
 
 const average = (arr) =>
   arr.reduce((acc, cur, i, arr) => acc + cur / arr.length, 0);
-const key = "59ec7bc1";
 
 export default function Index() {
-  const [movies, setMovies] = useState([]);
-
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState("");
   const [query, setQuery] = useState("game");
   const [selectedId, setSelectedId] = useState("");
+  const { movies, isLoading, error } = useMovies(query);
   /* const [watched, setWatched] = useState([]); */
+
   const [watched, setWatched] = useState(function () {
     let store = localStorage.getItem("watched");
     return JSON.parse(store);
@@ -94,51 +92,7 @@ export default function Index() {
         console.log("closing the movie screen");
       }
     });
-  }),
-    [];
-  useEffect(
-    function () {
-      const controller = new AbortController();
-      async function FetchMovie() {
-        try {
-          setIsLoading(true);
-          setError("");
-          const res = await fetch(
-            `http://www.omdbapi.com/?apikey=${key}&s=${query}`,
-            { signal: controller.signal }
-          );
-
-          if (!res.ok)
-            throw new Error("Error occured when the fetching is done");
-
-          const data = await res.json();
-          if (data.Response === "False") throw new Error("Movie not found");
-          console.log(data.Search);
-
-          setMovies(data.Search);
-
-          /* altenative;
-      fetch(`http://www.omdbapi.com/?apikey=${key}&s=game`).then(res=>res.json()).then(data=>setMovies(data.Search))
- */
-        } catch (err) {
-          if (err.name !== "AbortError") setError(err.message);
-        } finally {
-          setIsLoading(false);
-        }
-      }
-      if (!query.length) {
-        setMovies([]);
-        setError("");
-        setIsLoading(false);
-        return;
-      }
-      FetchMovie();
-      return function () {
-        controller.abort();
-      };
-    },
-    [query]
-  );
+  });
 
   return (
     <>
